Move more functions into visualiser
