+++ recognition/45801150_Task6/src/driver.py
@@ -26,7 +26,7 @@ def main():
     VQVAE.train_vqvae(vqvae, x_train_normalised, x_val_normalised, 2)
 
     # Test VQ-VAE performance on test set
-    test_images, reconstructed = VQVAE.compare_reconstructions(vqvae, x_test_normalised, 10)
+    test_images, reconstructed = visualiser.compare_reconstructions(vqvae, x_test_normalised, 10)
     visualiser.show_reconstructions(10, test_images, reconstructed)
 
     # Create and train PixelCNN
